Move Logger To Services
